server/llm/requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain_google_genai
langchain
langchain_community
python-dotenv
pypdf
langchain-huggingface
langchain_experimental
langgraph
psycopg[binary,pool]
tiktoken
langchain-postgres
langgraph-checkpoint-postgres
llama-index-core
llama-parse
llama-index-readers-file